He refactorizado el formulario de gastos y mejorado la interfaz de usuario:
- He mejorado la interacción del campo de fecha para facilitar su uso.
- He refactorizado la estructura de componentes de Finanzas: la lógica y JSX del formulario de gastos se movieron directamente a `src/components/finanzas/Finanzas.jsx`, y se eliminó `FormularioGasto.jsx`.
- He ajustado el diseño del formulario en `Finanzas.jsx` para que ocupe el 50% del ancho y esté centrado en vistas de escritorio (md y superiores) usando Tailwind CSS.
